labellerr-semantic-chatbot/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
pydantic-settings==2.3.4
sentence-transformers==3.0.1
numpy==1.26.4
chromadb==0.5.5
qdrant-client==1.9.2
google-generativeai==0.7.2
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.5
